Added some bonus and wrote docstrings

+++ Libft/ft_strlcat.c
@@ -6,18 +6,32 @@
 /*   By: muhakhan <muhakhan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/15 17:31:24 by muhakhan          #+#    #+#             */
-/*   Updated: 2024/11/14 17:56:06 by muhakhan         ###   ########.fr       */
+/*   Updated: 2024/12/03 20:16:07 by muhakhan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <unistd.h>
-#include <string.h>
+#include "libft.h"
 
-unsigned int	ft_strlcat(char *dest, const char *src, size_t size)
+/**
+ * ft_strlcat - Concatenates two strings ensuring null-termination.
+ * @dest: The destination string to which the source string is appended.
+ * @src: The source string to be appended to the destination string.
+ * @size: The total size of the destination buffer.
+ *
+ * This function appends the null-terminated string src to the end of dest.
+ * It will append at most size - strlen(dest) - 1 bytes, null-terminating
+ * the result. If the size is less than or equal to the length of dest,
+ * it returns size + strlen(src). Otherwise, it returns the length of the
+ * string it tried to create, i.e., the initial length of dest plus the
+ * length of src.
+ *
+ * Return: The total length of the string it tried to create.
+ */
+size_t	ft_strlcat(char *dest, const char *src, size_t size)
 {
-	unsigned int	j;
-	unsigned int	dest_len;
-	unsigned int	src_len;
+	size_t	j;
+	size_t	dest_len;
+	size_t	src_len;
 
 	j = 0;
 	dest_len = 0;
@@ -42,7 +56,7 @@ unsigned int	ft_strlcat(char *dest, const char *src, size_t size)
 // int main(void)
 // {
 // 	char dest[20];
-// 	unsigned int result;
+// 	size_t result;
 // 	memset(dest, 'r', 15);
 // 	result = ft_strlcat(dest, "lorem ipsum dolor sit amet", 5);
 // 	printf("Resulting string: %s\n", dest);
