+++ Libft/ft_strlcpy.c
@@ -6,17 +6,28 @@
 /*   By: muhakhan <muhakhan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/10 18:59:21 by muhakhan          #+#    #+#             */
-/*   Updated: 2024/11/14 17:56:10 by muhakhan         ###   ########.fr       */
+/*   Updated: 2024/12/03 20:16:19 by muhakhan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <string.h>
 #include <stdio.h>
 
-unsigned int	ft_strlcpy(char *dest, const char *src, size_t size)
+/**
+ * @brief Copies up to size - 1 characters from the NUL-terminated string src 
+ * to dest, NUL-terminating the result.
+ * 
+ * @param dest The destination buffer where the content is to be copied.
+ * @param src The source string from which the content is to be copied.
+ * @param size The size of the destination buffer.
+ * 
+ * @return The total length of the string it tried to create, i.e., the length 
+ * of src.
+ */
+size_t	ft_strlcpy(char *dest, const char *src, size_t size)
 {
-	unsigned int	i;
-	unsigned int	src_len;
+	size_t	i;
+	size_t	src_len;
 
 	i = 0;
 	src_len = 0;
